Initialise recurrence task class
